refactor: InputView에서 입력 검증을 완전히 처리하도록 개선
- InputView가 모든 입력 검증 담당
- 패키지 구조 재구성 (domain, view, util)
- .gitkeep 파일 제거"
