+++ src/main/java/lotto/domain/Lotto.java
@@ -19,7 +19,9 @@ public class Lotto {
 
     public Lotto(List<Integer> numbers) {
         validateNumbers(numbers);
-        this.numbers = numbers;
+        this.numbers = numbers.stream()
+                .sorted()
+                .toList();
     }
 
     public List<Integer> getNumbers() {
