+++ src/main/java/lotto/view/InputView.java
@@ -36,15 +36,28 @@ public class InputView {
         }
     }
 
-    public static int readBonusNumber() {
+    public static int readBonusNumber(List<Integer> winningNumbers) {
         System.out.println();
         System.out.println("보너스 번호를 입력해 주세요.");
         while (true) {
             try {
-                return Integer.parseInt(Console.readLine());
+                int bonusNumber = Integer.parseInt(Console.readLine());
+
+                validateBonusNumber(bonusNumber, winningNumbers);
+                return bonusNumber;
             } catch (NumberFormatException e) {
                 System.out.println("[ERROR] 보너스 번호는 숫자여야 합니다.");
+            } catch (IllegalArgumentException e) {
+                System.out.println(e.getMessage());
             }
+        }
+    }
+    private static void validateBonusNumber(int bonusNumber, List<Integer> winningNumbers) {
+        if (bonusNumber < 1 || bonusNumber > 45) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 1~45 사이여야 합니다.");
+        }
+        if (winningNumbers.contains(bonusNumber)) {
+            throw new IllegalArgumentException("[ERROR] 보너스 번호는 당첨 번호와 중복될 수 없습니다.");
         }
     }
 }
