+++ src/test/java/lotto/util/ValidatorTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.util;
 
-import lotto.util.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
