+++ src/test/java/lotto/view/InputViewTest.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
+import lotto.Application;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
